chado/requirements.txt
requests
python-dotenv
chromadb
streamlit
pypdf
python-docx
pyyaml
sentence-transformers
torch
